refactor(runtime): reorganize runtime into game_loop, render, and frame subpackages
