+++ docs/research/render_loop_frame_pacing.md
@@ -10,15 +10,15 @@ has spare capacity.
 ## Materials
 
 - `src/heart/runtime/rendering/timing.py`
-- `src/heart/runtime/render_planner.py`
-- `src/heart/runtime/game_loop.py`
+- `src/heart/runtime/render/planner.py`
+- `src/heart/runtime/game_loop/core.py`
 - `docs/library/runtime_systems.md`
 
 ## Source excerpts
 
 - "`self.average_ms = (ema_alpha * duration_ms) + ((1.0 - ema_alpha) * self.average_ms)`" (`src/heart/runtime/rendering/timing.py`)
-- "`estimated_cost_ms, has_samples = self._timing_tracker.estimate_total_ms(`" (`src/heart/runtime/render_planner.py`)
-- "`clock.tick(self.max_fps)`" (`src/heart/runtime/game_loop.py`)
+- "`estimated_cost_ms, has_samples = self._timing_tracker.estimate_total_ms(`" (`src/heart/runtime/render/planner.py`)
+- "`clock.tick(self.max_fps)`" (`src/heart/runtime/game_loop/core.py`)
 
 ## Notes
