+++ docs/research/render_pipeline_refactor.md
@@ -26,6 +26,6 @@ simpler to evolve per-renderer behavior without touching dispatch logic.
 
 ## Materials
 
-- `src/heart/runtime/render_pipeline.py`
+- `src/heart/runtime/render/pipeline.py`
 - `src/heart/runtime/rendering/renderer_processor.py`
 - `src/heart/runtime/rendering/timing.py`
