+++ src/heart/renderers/mandelbrot/scene.py
@@ -676,7 +676,7 @@ def main() -> None:
 
     from heart.device.local import LocalScreen
     from heart.runtime.container import build_runtime_container
-    from heart.runtime.render_pipeline import RendererVariant
+    from heart.runtime.render.pipeline import RendererVariant
     from heart.utilities.env import Configuration
 
     profiling = os.environ.get("PROFILING", "False").lower() == "true"
