+++ src/heart/runtime/game_loop/__init__.py
@@ -1,0 +1,4 @@
+"""Runtime game loop orchestration."""
+
+from .core import GameLoop as GameLoop
+from .core import RendererVariant as RendererVariant
